[FEAT] #1 - 파일 분리
