+++ Instagram/lib/feed.dart
@@ -1,0 +1,81 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class Feed extends StatelessWidget {
+  const Feed({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start, // 텍스트 왼쪽 정렬
+      children: [
+        // 이미지
+        Image.network(
+          "https://cdn2.thecatapi.com/images/kat_7kqBi.png",
+          height: 400,
+          width: double.infinity, // 부모에서 넣을 수 있는 최대 크기
+          fit: BoxFit.cover, // scaleType과 비슷
+        ),
+        // 아이콘
+        Row(
+          children: [
+            IconButton(
+              onPressed: () {},
+              icon: Icon(
+                CupertinoIcons.heart,
+                color: Colors.black,
+              ),
+            ),
+            IconButton(
+              onPressed: () {},
+              icon: Icon(
+                CupertinoIcons.chat_bubble,
+                color: Colors.black,
+              ),
+            ),
+            Spacer(), // 빈 공간 추가 - 커질 수 있을만큼 최대한 커짐
+            IconButton(
+              onPressed: () {},
+              icon: Icon(
+                CupertinoIcons.bookmark,
+                color: Colors.black,
+              ),
+            ),
+          ],
+        ),
+
+        // 좋아요
+        Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Text(
+            "2 likes",
+            style: TextStyle(
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+        ),
+
+        // 설명
+        Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Text(
+            "My cat is docile even when bathed. I put a duck on his head in the wick and he's staring at me. Isn't it so cute??",
+          ),
+        ),
+
+        // 날짜
+        Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Text(
+            "FEBURARY 6",
+            style: TextStyle(
+              color: Colors.grey,
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
