+++ Instagram/lib/home_page.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:instagram/feed.dart';
+
+class Homepage extends StatelessWidget {
+  const Homepage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        leading: IconButton(
+          icon: Icon(CupertinoIcons.camera, color: Colors.black),
+          onPressed: () {},
+        ),
+        // actions에는 여러 개의 원소가 가능하기 때문에 배열 []
+        actions: [
+          IconButton(
+            icon: Icon(CupertinoIcons.paperplane, color: Colors.black),
+            onPressed: () {},
+          )
+        ],
+        title: Image.asset(
+          'assets/logo.png',
+          height: 32,
+        ),
+        centerTitle: true,
+        backgroundColor: Colors.white,
+      ),
+      body: Feed(),
+    );
+  }
+}
