+++ Instagram/lib/main.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'home_page.dart';
 
 void main() {
   HttpOverrides.global = NoCheckCertificateHttpOverrides(); // 생성된 HttpOverrides 객체 등록
@@ -17,104 +17,6 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       home: Homepage(), // home에 화면에서 보여줄 첫 번째 위젯을 넣음
     );
-  }
-}
-
-class Homepage extends StatelessWidget {
-  const Homepage({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-   return Scaffold(
-     appBar: AppBar(
-       leading: IconButton(
-         icon: Icon(CupertinoIcons.camera, color: Colors.black),
-         onPressed: () {},
-       ),
-       // actions에는 여러 개의 원소가 가능하기 때문에 배열 []
-       actions: [
-         IconButton(
-           icon: Icon(CupertinoIcons.paperplane, color: Colors.black),
-           onPressed: () {},
-         )
-       ],
-       title: Image.asset(
-         'assets/logo.png',
-         height: 32,
-       ),
-       centerTitle: true,
-       backgroundColor: Colors.white,
-     ),
-     body: Column(
-       crossAxisAlignment: CrossAxisAlignment.start, // 텍스트 왼쪽 정렬
-       children: [
-         // 이미지
-         Image.network(
-           "https://cdn2.thecatapi.com/images/kat_7kqBi.png",
-           height: 400,
-           width: double.infinity, // 부모에서 넣을 수 있는 최대 크기
-           fit: BoxFit.cover, // scaleType과 비슷
-         ),
-         // 아이콘
-         Row(
-           children: [
-             IconButton(
-               onPressed: () {},
-               icon: Icon(
-                 CupertinoIcons.heart,
-                 color: Colors.black,
-               ),
-             ),
-             IconButton(
-               onPressed: () {},
-               icon: Icon(
-                 CupertinoIcons.chat_bubble,
-                 color: Colors.black,
-               ),
-             ),
-             Spacer(), // 빈 공간 추가 - 커질 수 있을만큼 최대한 커짐
-             IconButton(
-               onPressed: () {},
-               icon: Icon(
-                 CupertinoIcons.bookmark,
-                 color: Colors.black,
-               ),
-             ),
-           ],
-         ),
-
-         // 좋아요
-         Padding(
-           padding: const EdgeInsets.all(8.0),
-           child: Text(
-             "2 likes",
-             style: TextStyle(
-               fontWeight: FontWeight.bold,
-             ),
-           ),
-         ),
-
-         // 설명
-         Padding(
-           padding: const EdgeInsets.all(8.0),
-           child: Text(
-             "My cat is docile even when bathed. I put a duck on his head in the wick and he's staring at me. Isn't it so cute??",
-           ),
-         ),
-
-         // 날짜
-         Padding(
-           padding: const EdgeInsets.all(8.0),
-           child: Text(
-             "FEBURARY 6",
-             style: TextStyle(
-               color: Colors.grey,
-             ),
-           ),
-         ),
-       ],
-     ),
-   );
   }
 }
